# lib/api-zod/src/generated/types/mediaCreateRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface MediaCreateRequest {
  title: string;
  description?: string | null;
  link?: string | null;
  imageUrl?: string | null;
  secondaryImageUrl?: string | null;
  categoryId?: string | null;
  categoryName?: string | null;
  mainCategoryId?: string | null;
  mainCategoryName?: string | null;
  subCategoryId?: string | null;
  subCategoryName?: string | null;
  playStoreLink?: string | null;
  appStoreLink?: string | null;
  figmaLink?: string | null;
  websiteLink?: string | null;
}