# lib/api-zod/src/generated/types/project.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Project {
  id: string;
  title: string;
  description?: string | null;
  link?: string | null;
  imageUrl?: string | null;
  previewImageUrl?: string | null;
  categoryId?: string | null;
  categoryName?: string | null;
  mainCategoryId?: string | null;
  mainCategoryName?: string | null;
  subCategoryId?: string | null;
  subCategoryName?: string | null;
  sortOrder: number;
  isHidden: boolean;
  lastCheckedAt?: number | null;
  createdAt: number;
  updatedAt?: number | null;
}